Move local server API into extension
To use the API are only implemented in local server, create an
extension that register the 'local' namespace to use these API.
