+++ utils/tangleid/handleErrors.js
@@ -1,0 +1,8 @@
+function handleErrors(response) {
+  if (!response.ok) {
+    throw Error(response.statusText);
+  }
+  return response;
+}
+
+module.exports = handleErrors;
